LayoutTests/fast/mediastream/enumerateDevices-microphone-denied-expose-with-capture.html

<!doctype html>
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
promise_test(async (test) => {
    const stream = await navigator.mediaDevices.getUserMedia({ video : true });
    test.add_cleanup(() => stream.getTracks().forEach(t => t.stop()));

    if (window.testRunner)
       testRunner.resetUserMediaPermission();
    const devices = await navigator.mediaDevices.enumerateDevices();
    for (const device of devices)
        assert_true(device.label === "" || device.kind !== "audioinput");
}, "enumerateDevices() filters out microphones if microphone permission is prompt");

promise_test(async (test) => {
    if (window.testRunner)
        testRunner.setMicrophonePermission(false);

    const stream = await navigator.mediaDevices.getUserMedia({ video : true });
    test.add_cleanup(() => stream.getTracks().forEach(t => t.stop()));
    const devices = await navigator.mediaDevices.enumerateDevices();
    for (const device of devices)
        assert_true(device.label === "" || device.kind !== "audioinput");
}, "enumerateDevices() filters out microphones if microphone permission is denied");

promise_test(async (test) => {
    if (window.testRunner)
        testRunner.setMicrophonePermission(true);

    await navigator.mediaDevices.getUserMedia({ video : true });
    const devices = await navigator.mediaDevices.enumerateDevices();
    for (const device of devices)
        assert_true(device.label === "" || device.kind !== "audioinput");
}, "enumerateDevices() filters out microphones even if microphone permission is granted and camera is in use");

    </script>
</head>
<body>
</body>
</html>
